Changes - See desription for details
Created 2 scripts for analyzing LAS/tether drag experimental data for System ID work.
Created a "Weight" vehicle variant for simulation of the tether drag system ID work
Revised the LAS simulink model to account for the tangency condition of the tether as it exits the LAS.
Added a property to the LAS class for the polynomial fit values for the tether exit elevation angle and the LAS elevation angle.
Added the CB/CG change to the manta build script. Rebuilt the manta vehicle to reflect changes.
Changed the manta path following controller to output ctrlSurfDefl instead of ctrlSurfDeflection to match majority of plant models.
Added the weight variant to the vehicle simulink model.
Removed the scope from the realistic sensors model.
Removed the portion of the LAS drag block diagram which calculated the normal component of force. This is accounted for in the dr/dtheta block

-Andrew